fixing up changed code. dotnet build now down to 0 errors, 0 warnings
