+++ src/Nuget.config
@@ -10,7 +10,7 @@
     <add key="dotnet.core" value="https://dotnet.myget.org/F/dotnet-core/api/v3/index.json" />
     <add key="RC2" value="https://www.myget.org/F/aspnetvnext/api/v3/index.json" />
     <add key="xunit" value="https://www.myget.org/F/xunit/" />
-    <add key="NpgsqlUnstable" value="https://www.myget.org/F/npgsql-unstable/api/v3/index.json" />
+     <add key="NuGet" value="https://api.nuget.org/v3/index.json" />
 
 
   </packageSources>
