+++ src/project.json
@@ -18,7 +18,12 @@
         "Microsoft.Extensions.Logging": "1.0.0-rc2-final",
         "Microsoft.Extensions.Logging.Console": "1.0.0-rc2-final",
         "Microsoft.Extensions.Logging.Debug": "1.0.0-rc2-final",
-        "Microsoft.EntityFrameworkCore.Sqlite": "1.0.0-rc2-final"
+        //"Microsoft.EntityFrameworkCore.Sqlite": "1.0.0-rc2-final"
+       "Npgsql.EntityFrameworkCore.PostgreSQL": "1.0.0-*",
+       "Microsoft.EntityFrameworkCore": "1.0.0-rc2-final",
+        "Microsoft.Extensions.Configuration.Json": "1.0.0-rc2-final",
+          "Microsoft.Extensions.Configuration.EnvironmentVariables": "1.0.0-rc2-final"
+
     },
     "tools": {
         "Microsoft.AspNetCore.Server.IISIntegration.Tools": {
